Revert "Splitting the chapter on etale cohomology"
This reverts commit 2ada00ef856ad601a2495ca8b2657ee3878ddecf.

We decided to have a separate branch where we work on splitting the
chapter on etale cohomology, so we can do some experiments with python
scripts etc, without being too worried about destroying previous work.
Hence in the master branch we leave the chapter on etale cohomology as a
single chapter for the moment.