Weil restriction of closed subscheme along flat pure map
I propose to add a lemma which is a corollary to theorem tag=05PF
(currently thm 33.23.3) to the representability of the Weil restriction
of closed subschemes. This is important for applications to
representability of equalizers, kernels (etc.) like in SGA3, tome 2,
Exp. VIII, thm 6.4. (See also prop. B.3 in Abramovich, Romagny, "Moduli
of Galois p-covers in mixed characteristics".)