Question 04PZ
Hi,

I've attached what seems to be an affirmative answer to the question posed in
Remark 04PZ in the case of a finite extension. I have no idea about the
general case, but as the remark explicitly mentions the finite case, I
thought it was still worth writing it down.

Best,
Philipp