Skeleton section about strict transform
A tricky point is that the strict transform doesn't just depend on the
morphism, but also on the center of the blow up. More precisely, it
depends on the underlying closed subset of the chosen center. To split
hairs even more: If two centers Z_1, Z_2 give isomorphic blowups and
|Z_1| = |Z_2|, then strict transforms constructed with them agree.