Remove files which have found a home in stacks-website
Also adapted the Makefile to account for the changes:
(1) The "tags" target is simplified as it no longer has to take
care of php and html files (which have been removed)
(2) The tags_install target now just copies the dvis, pdfs, tar,
and the actual files comprising the stacks project to the install
directory. Finally, it adds a VERSION file.