Alter 'make tags' target
With the new website setup we are compiling the live versions of
the chapters and book in place. We change this by placing a
version of scripts/latex.sh in tags/ and using this to compile
the chapters and book.

An added benefit of this is that the script allows us to use
parallel make for the make tags target. This was previously
pointed out on the blog here:

http://math.columbia.edu/~dejong/wordpress/?p=651