More on limits of algebraic spaces
This material is on the one hand tremendously boring and
on the other hand, quite tricky. It turns out for example
that it is easy to show that given a filtered inverse system
of qc+qs schemes with separated limit, then the objects
are separated for large indices. This we missed the first
time through... Anyway, the goal of working through this
material with the current level of detail is to be able
to mirror most of the arguments for proving Chow's lemma
in the case of algebraic spaces...

We'll see.