Finiteness cohomology coherent sheaves proper spaces
No surprises. Decided to avoid using clever spectral sequence argument
and instead use vanishing of higher direct images for a high power
of a relatively ample sheaf as in the schemes case.

The discussion at the end of the proof of why the scheme theoretic
support of the sheaf G is all of Z is perhaps a bit too detailed.