Clean up the material from last commit
The claims on duals of pseudo-coherent complexes were too optimistic. As
a replacement we directly construct a base change map for pushforwards
of internal homs. Then we show this is a qis by the induction principle.

We also fixed the proof of the lemma on duals of perfect objects as the
previous one was nonsense.

Finally, we improve the (trivial) lemma on vanishing of Ext groups
between objects of the derived category.