Revised chapters "Examples", "Descent", "Algebraic Spaces"
I attach a proposed revised version of the chapter "Examples"
from the Stacks Project. I added another example by Hironaka,
showing that projective morphisms do not satisfy descent
in the fpqc topology on schemes. I modified the earlier example
by Hironaka (showing that projectivity is not Zariski-local)
to match the notation in this new example.

I also attach revised versions of the chapters "Descent"
and "Algebraic Spaces", with brief links to the new example.