Stop parse.py from complaining too much
-- the \section{...} command had better be on one line because I am not
sure what the consequences will be for the scripts used to munge the
stacks project data if the title spans two lines. As a consequence we
now have a line in examples.tex which is more than 80 characters long.
As far as I know none of the python or php scripts actually depend on
having 80 characters per line. (I couldn't think of a suitable shorter
title of this section.)
-- We require $$ to be alone by themselves on a line.
-- Fixed a reference.