Fix error in artin.tex pointed out by David Rydh
The proof of Lemma Tag 07YM was wrong in two ways:

-- We were using the condition (RS*) in some cases where it didn't
   apply, namely for an extension A' of a Noetherian ring A by a
   module isomorphic to the residue field of a finite type point.
-- The second mistake was in some sense the same mistake thinking that
   the ring A' was Noetherian.

The solution, as pointed out by David Rydh, is twofold:

-- Strengthen assumption (RS*) to cover the first snafu
-- Only prove openess of versality at closed points

As far as I can see, this is harmless in all applications.

There are still some things left to fix. In particular, some of
the material concerning T_x(-) has to be fixed to deal with
non-finitely generated modules.