Excise finite type conditions
We need to have the functors T_x and Inf_x for all modules not just
for finite type ones. This actually simplifies the discussion a
tiny bit as we don't have to carry along the finite type conditions
everywhere.

Thanks to David Rydh for pointing this out (see
commit 2ccbbe3087e4dc2b1df2193c81ede7486931424c for more
information).