Descent for universally injective morphisms
A while back, I mentioned that there is a natural way to describe
exactly which ring homomorphisms R -> S admit descent for modules: they
are the ones which are universally injective. That is, for any R-module
M, M -> M tensor_R S should be injective.

I typed up some notes about this; actually I did this some months ago
and meant to polish them some more, but never got around to it.

If this has a natural place in the Stacks Project, you're welcome to
incorporate it. Otherwise, it'll probably jut rot on my web site...

Kiran