Hypotheses of Lemma 089E
Dear Johan,

I was wondering: in the lemma in question, could one weaken the
assumptions on g by requiring that it be a quasi-compact schematic
(representable by schemes) morphism of algebraic spaces? The present
assumptions imply that Y is a scheme already which perhaps was not
intended.

Best,
Kestutis

Answer: Yes! Thanks!

In this commit we do what is asked above and we modify the proof
accordingly.