Taking Rlim sometimes gives nice complexes
Three results. The final one tells us we can obtain perfect complexes
over I-adically complete rings as limits of perfect complexes over the
quotients A/I^n.

This kind of thing is used in crystalline cohomology.

Question: is there a nice way to use the results on derived completion
in the setting of crystalline cohomology?