Final commit on stack of coherent sheaves on algebraic spaces
OK, the situation is now as follows:
(0) Philosofically speaking the result holds because you can prove it
using methods from derived algebraic geometry.
(1) It is clear that the method as presented works.
(2) Quite a few details are omitted, mostly of a nature of checking
that certain diagrams commute -- some of these checks have been done in
the literature (eg Illusie) and others are variants of these.

However, at this point it seems preferable to continue on with
applications to Quot spaces for example rather than continuing to
add innumerable details.