Chapter on fields from CRing project
This is to test if it is feasable to have a chapter on fields which
comes before the chapter on commutative algebra. The idea being that
people with an undergraduate algebra course should be able to read it.

The text of the chapter is take from the CRing project with permission
from Akhil Mathew. Contributions to this chapter in particular from:
Akhil Mathew, Geoffrey Lee, and Ryan Reich