Fix a mistake
The error was in thinking that the fibre of an lci map is lci which is
true only when the map is flat. Fixed.

Also added a section on complete intersection maps of Noetherian local
rings. This is not as nice as the absolute notion because in general the
map from a Noetherian local ring to its completion is not a complete
intersection. Still it is useful in order to state the most general
version of Avramov's result and then to relate it to the material about
complete intersection homomorphisms in the finite presentation case
which is discussed earlier.