LaTeX
Introduced a macro

\def\QCoh{{\textit{QCoh}}}

and replaced almost all occurences of \textit{QCoh} by \QCoh. There are
still some places where we use \textit{QCoh}, namely in the chapter
on examples of stacks. This is because the meaning there is different;
it indicates a category fibred in groupoids over Sch and not the notational
gadget that takes a ringed topos and spits out its category of
quasi-coherent modules.