New condition on cocontinuous functors
In this commit we introduce an additional property a cocontinuous
functor u : C ---> D between sites could have. The property is roughly
that one can ``pull back'' coverings to coverings along u. This
property is often satisfied in practice (for example localization
functors always satisfy it) and had so far escaped our attention. It is
useful in the setting of simplicial sites defined using cocontinuous
functors (see next commit).