Partial fix of error found by Bhargav Bhatt
The mistake was that the functor v of the first lemma of the commit is
not continuous although it does send coverings to coverings. Thus the
applications of various lemmas from the chapter on sites aren't allowed
and we need another arguement to see that i^{-1} has a left adjoint.

First of all, by abstract nonsense, there is going to be such an adjoint
as soon as i^{-1} commutes with all limits. And, as Bhargav and Peter
point out this is true because we may check this on weakly contractible
objects V over Z where the formula i^{-1}F(V) = F(vV) is correct.

On the other hand, we already have a machine for making such an adjoint
in the abelian categories case, namely

http://stacks.math.columbia.edu/tag/0793

which we have used already once, namely in

http://stacks.math.columbia.edu/tag/0797

to construct a lower shriek functor. Thus fixing the second lemma is
just a matter of translating this from abelian categories to topoi which
should be no problem. Stay tuned...