Redefine valuation ring
It is MUCH better to allow fields to be valuation rings. In fact, some
of the arguments on specializations of points and valuation rings mapping
to schemes were wrong, or at least very confusing with the definition as
given originally.

Thanks to Brian Conrad!

Here are his reasons for making this change (if there are any mistakes
in what follows, then I take full responsibility, but in any case the
point we are trying to make here is *only* that we really should change
the definition in the Stacks project).

"Johan, below are some reasons that I think it is good to include fields
as examples of valuation rings (which isn't to say that fields are
considered to be Dedekind domains; that is not a convention which I am
advocating). In effect, it amounts to a bunch of reasons for not
excluding the trivial valuation on a field.

Does the Stacks Project rule out fields from being valuation rings?

1.  For the definition of Riemann-Zariski spaces RZ(K/k) relative to an
extension of fields K/k and the proof of
their quasi-compactness, it is very natural that we don't omit the
trivial valuation.  The same holds for when trying to relate
valuation rings contained in a given valuation ring by lifting
valuations on the residue field.

2. The integral closure of any domain D is the intersection of the
valuation rings on its fraction field that contain the domain...except
that if we don't consider fields to be valuation rings then this theorem
would not be true in the (admittedly trivial) case that D is itself a
field.

3. The definition of local ring should not exclude the case of fields
(it would wreak havoc with the notions of local scheme, locally ringed
space, etc.), so we'd like to say that a local ring is a valuation ring
iff it is maximal with respect to domination among local domains with a
common fraction field.  But we would have to exclude from this the case
of local rings which are fields if we don't regard fields as valuation
rings.

4. We'd like to say that any local subring of a field K is dominated by
a valuation ring having fraction field K.  This is false for various
cases with local subrings that are fields if we do not allow valuation
rings to be fields.

5. The various valuative criteria are true (though not so useful) when
allowing valuation rings that are fields.

6. Does the trivial valuation on a field not have a "valuation ring" (in
contrast with all other non-archimedean valuations on a field)?

7. Trivial valuations play a useful technical role in the middle of some
proofs when setting up the theory of adic spaces (even though the v_x's
are not going to be trivial for adic spaces arising from rigid spaces
over non-archimedean fields). In a related matter, the fibers of the map
Spa(A,A^0) ---> M(A) from an adic k-affinoid space to the corresponding
Berkovich space are something like Riemann-Zariski spaces (if I remember
correctly), so in all of the fibers there should be points corresponding
to trivial valuations on residue fields.  For instance, with the adic
closed unit disc the fiber over each type-2 point has fiber that's like
RZ(\kappa(T)/\kappa) for various fields \kappa."