LaTeX environment for dealing with references
This environment is called "reference". We currently use the comment
environment from the verbatim package to ignore anything in between
\begin{reference} and \end{reference} when running latex.

Coding style. We expect this environment to be used as follows

\begin{lemma}
\label{lemma-ABC}
\begin{reference} \cite[page 13]{DEF} \end{reference}
....
\end{lemma}

Or in case there are multiple references for a single result or some
explanation is warranted, we use it like so

\begin{lemma}
\label{lemma-ABC}
\begin{reference}
The first half of this lemma can be found in \cite[page 13]{DEF} and the
second half in \cite[III, Lemma 14.2.3]{EGA}
\end{reference}
....
\end{lemma}