Stein factorization for morphisms of algebraic spaces
Actually, the proof of the (general) schemes case had a slight snafu
which we fix in this commit as well. Moreover, it actually is more
efficient the way we do things now, at the cost of adding some more
results on limits of schemes. It seems you can never have enoug of
these types of results; to me it seems results arguments with limits
of schemes are like the epsilon-delta arguments in analysis.