Fix an error
While reviewing the material on cohomology and base change we found
a mistake in the proof of Lemma lemma-base-change-RHom. The mistake is
fatal in the sense that we don't know if the result is true as it was
stated. The fix is to assume that the object E is a perfect complex (as
opposed to just pseudo-coherent).

We also reorganized the material somewhat. We intend to add a fifth
section on cohomology and base change in the near future (removing some
of the Noetherian hypotheses by limit arguments).