Essentially constant in additive categories
Here we show that if $F, G : \mathcal{I} \to \mathcal{A}$ are functors
from a filtered category to an additive category, then $F \oplus G$ is
essentially constant if and oly if $F$ and $G$ are essentially constant,
provided that $\mathcal{A}$ is Karoubian.

If $\mathcal{A}$ is not Karoubian, then we cannot even start the
"obvious" proof as we don't know that \colim (F \oplus G)$ can be split
into two parts! Presumably this result is actually false if
$\mathcal{A}$ is not Karoubian.

Thanks to Nuno Cardoso for pointing out this difficulty (see also next
commit).