Valuative criteria
Thanks to Brian Conrad

Here is a part of his email concerning the topic of this commit:

"Here is a more direct way to say what is going on in the case of alg.
spaces, it case it might be of some use to include a Remark along such
lines in the Stacks Project.  Let f:X ---> Y be a quasi-compact
separated map between quasi-separated alg. spaces.  Let R be a valuation
ring with fraction field k, and suppose we are given y in Y(R) and x_k
in X(k) over the associated y_k in Y(k).   We want to consider the
problem of whether x_k extends uniquely to an x in X(R) over y, and
possibly after some local extension on R to a bigger valuation ring.  We
can at least pull back along y so that we may rename Y as Spec(R).  That
is, we're give X = qc separated algebraic space over Spec(R), and x_k in
X(k).  We wonder if it extends to X(R), possibly after some local
extension on R to a bigger valuation ring.  Since X_k is separated, so
x_k is a closed immersion into X_k, there is no harm in replacing X with
the "schematic closure" of x_k.

This reduces our study to when X_k = Spec(k) and X is R-flat (as
flatness over val. ring is the same as being torsion-free).  In such a
situation, the key thing is to show that X is univ. closed over Spec(R)
iff X = Spec(R).  The implication "<==" is obvious, and for the converse
it suffices to show X is quasi-finite over Spec(R) (as then X is a
*scheme*, so we can apply the usual thing).  To check being quasi-finite
it is harmless to make a local extension on R to a bigger valuation ring
since that is an fpqc base change (and such base change preserves the
hypotheses we have arranged to hold).  But if we can make such a base
change to acquire a section then the section is a closed immersion (as X
is separated) and its defining ideal must vanish (since by R-flatness
this can be checked at the generic point, where all is clear)."