A little bit more about dualizing complexes
Existence implies universally catenary and finite dimension for rings

Definition for locally Noetherian schemes... We just say it should be
affine locally a dualizing complex. This is not the most wonderful way
to do it, but it avoids talking about finite injective dimensions for
complexes of modules over a sheaf of rings.