Expand the section on local cohomology
The previous version of this section was written for the application of
it in the chapter on the pro-etale site. Hence the point of view was to
stress how one can compute the derived functor R\Gamma_I of taking
I-power torsion over Noetherian rings, by the extended alternating Cech
complex.

However, for the general development, taking the approach with the
alternating Cech complex is the correct one, because it gives us the
right adjoint R\Gamma_Z to the inclusion functor

D_{I-power torsion}(A) --------> D(A)

which we don't get by taking R\Gamma_I. Moreover, we match R\Gamma_Z
with taking cohomology supported in Z on the corresponding affine
scheme.