Equivelence beteen torsion and derived complete modules
Wonderful!

Thanks to Bhargav Bhatt for pointing out that this is helpful for the
formulation of Grothendieck's duality statement in local cohomology
(coming up soon).

Many people use the torsion picture to think about derived complete
objects (for example in work of Dan Halperin-Leistner and Antoly
Preygel) and they know this is the same thing as the correspondence is
in Jacob Lurie's work somewhere. Another reference is

Dwyer, W. G. and Greenlees, J. P. C.
Complete modules and torsion modules

which appears to treat a more general question.