Terminology "quasi-isomorphic"
The phrase "K and L are quasi-isomorphic" is supposed to mean that K and L
are isomorphic as objects in D(R). But since we never formally defined
"quasi-isomorphic" there was no way for readers to know that.

The solution is either to stop using the phrase "quasi-isomorphic"
(but it gets used 92 times), or to make a formal definition. For the
moment I have just added a line to the chapter on Derived Categories
saying what the terminology means, but I have stopped short of actually
putting it in a definition, because now I am no longer so sure that it
is standard terminology... Please weigh in on this issue if you have
an opinion.

Thanks to Jyoti Prakash Saha
http://stacks.math.columbia.edu/tag/0657#comment-603
http://stacks.math.columbia.edu/tag/0657#comment-605