Finding invariant affine opens
Thanks to Angelo Vistoli for discussions

This somewhat long commit constructs invariant affine opens for
codimension 1 points on a finite groupoid. The method is a bit
convoluted. It is essentially a 2 step process where we first find a
dense open where the groupoid is flat and then we carefully glue on an
affine open of the codimension 1 points along the boundary.

We will apply this to prove that a Noetherian algebraic space is a
scheme away from codimension 2. To prove this one can simplify the
arguments in this commit a bit and we will put in a remark pointing this
out later on.