Dimension of a stack
I would like to add the notion of dimension of an algebraic stack
to the Stacks Project. At least, I can't find the definition there,
and I think it is really basic and convenient. Once the definition
is there, one could mention the dimension of various stacks
that come up as examples in the Stacks Project.

I attach a Latex file with a proposed write-up of the dimension of a
locally Noetherian stack, following Laumon and Moret-Bailly,
equations 11.14 (definition of dimension of a stack at a point)
and 11.15 (definition of dimension of a stack).
I had to prepare some ground about schemes and algebraic spaces first.
So the attached file
is intended to be divided into three parts: (1) add to the end of the
section "Smooth morphisms" in the chapter
"Morphisms of Schemes";
(2) add to the end of the section "Smooth morphisms" in the chapter
"Morphisms of Algebraic Spaces";
and (3) a new section for the chapter "Properties of Algebraic Stacks".

I tried to label the way my file is divided into three parts.

I hope you can use this for the Stacks Project.

Best wishes,
Burt