More material on the formal spectrum
Unfortunately, making things more explicit is also uglier.

Observation: in Yasuda and also to some extend in McQuillan they have a
notion of linearly topologized module which is different from ours. On
the other hand, ours is the same as the one in Bourbaki Commutative
Algebra: Chapters 1-7, Page 236, Exercise 14 and our (completed) tensor
product is the same as the one introduced in ibid. Exercise 28.