Comment about Tag 0315
In the proof of Tag 0315 the ideal I need not need to be nilpotent.
Instead we use that I/I^n is nilpotent in the ring R/I^n. Your
proof is essentially how the proof of Nakayama's lemma in the
nilpotent case works.

I have added a sentence with the explanation.

The reason for not adding your argument is that generally speaking
we try to avoid having the same thing proven in multiple locations,
although in this case your argument is so short that in some sense
it might be quicker to put it in.

On Tue, Sep 16, 2014 at 3:51 PM, Minseon Shin <shinms@berkeley.edu> wrote:
> Hello,
>
> Here is a comment regarding the beginning of the proof of Lemma 10.91.1 (Tag
> 0315). I think the reference to Nakayama's lemma (Lemma 10.18.1, Tag 00DV)
> is sort of confusing here because, unless I am mistaken, it seems to me that
> the claim "M/IM \to N/IN surjective implies M/I^{n}M \to N/I^{n}N surjective
> for all n \ge 1" is true even if M and N are not finitely generated and I is
> not nilpotent. Here is my claimed proof (which is by induction on n):
>
> It is equivalent to say "\phi(M)+IN = N implies \phi(M)+I^{n}N = N for all n
> \ge 1". Suppose \phi(M)+I^{n}N = N. Then I\phi(M)+I^{n+1}N = IN. Adding
> \phi(M) to both sides gives \phi(M)+I^{n+1}N = \phi(M)+(I\phi(M)+I^{n+1}N) =
> \phi(M)+IN = N.
>
> Best,
> Minseon