Mistake propagated from Raynaud-Gruson
Thanks to Ofer Gabber

The mistake was that there is in general no map

N \otimes_B B' ----> M \otimes_A A^h

extending the given map u : N ---> M. The current proof therefore only
works in case u is a map of B-modules, which is also the only case we
currently use in the Stacks project. However, we will ASAP add a proof
of the general case (due to Gabber).