Base change for ring maps essentially of finite type
TODO: Write about morphisms of schemes which are essentially of finite
type. There seem to be several different possible definitions, one of
which is to say that X ---> S is essentially of finite type if X is a
cofiltered limit of schemes of finite type over S with affine open
immersions as transition maps.