Minor corrections on algebra matters.
algebra.tex:355: The "this" in the "Informally this means..." paragraph
refers to the notion of "finitely presented", and so belongs inside part (2) of the
definition, rather than after it. (Otherwise it should be more explicit
about what it means by "this".)

algebra.tex:718: I have never seen transitive relations referred to as
"associative", so I take this for a typo.

algebra.tex:1863: There are various R-modules around, so I figured a bit more
explicitness wouldn't hurt.

algebra.tex:2021: It is certainly not the only iso $f : S^{-1}R \otimes_R M
\to S^{-1}M$.

algebra.tex:8067: The "(faithfully)" alternatives were added to the second
sentence to help prove the first sentence. It wouldn't hurt splitting the
lemma in two, though, as it now makes four different claims.

more-algebra.tex:4556: Fixed to the best of my knowledge. I am not 100% sure
of this since I don't know your notations.