Fix idiotic mistake about graded projective modules over dgas
Thanks to Rishi Vyas for pointing out the error:

"Lemma 22.26.1 is, unfortunately, not quite correct; it essentially says
that graded projective modules over $A$ are K-projective and this is
false.

The classical counterexample is over the ring of dual numbers
$k[\epsilon]$, where $k$ is your favorite field. The complex $\ldots
k[\epsilon] \xrightarrow{\epsilon} k[\epsilon] \xrightarrow{\epsilon}
k[\epsilon] \ldots$ is graded projective and acylic, but not
contractible.

I think the problem with the proof is in the line "The associated long
exact sequence ... ... thereby proving part (1)." - the first and last
lines in the diagram are necessarily isomorphisms. If, for example, you
take a non-contractible chain map from $P$ to $Z[1]$ which is the image
of a chain map from $P$ to $M$, then it must map to zero under the
boundary operator of the exact sequence.

There is also problem with the line "In particular, if $Z\subset M$ is
the kernel ... ... is acyclic": if $A$ is a DGA with non-zero
differentials, $Z$ is not necessarily a sub-module of $M$. However,
since the counterexample mentioned above is with $A$ a ring, this is not
the point where the primary obstruction to the veracity of the lemma
lies. "

Changes necessitated by this observation:

(1) Turn Lemma Tag 09R0 into Remark Tag 09R0 stating what is not true
(2) Turn Lemma Tag 09R1 into Remark Tag 09R1 asking what is true
(3) Weaken Proposition Tag 09R3 to remove one case of statement
(4) Modify the statement of Lemma Tag 09RB.

Only Lemma Tag 09RB was used outside of the Section in question and
luckily Lemma Tag 09RB was used only in one direction, namely, to show
that a compact object has a certain representative.

Also: updated tags file and added Rishi Vyas as a collaborator.