Improvements to intersection.tex
Thanks to Dmitry Korb for help simplifying the proof of 0B0W
Thanks to Ashwin Deopurkar finding a mistake in the proof of 0B1V

Other changes: work in a bit greater generality, especially in the
earlier lemmas. This is useful, because it allows us to work locally on
X in order to prove equalities of cycles.

Also added: a remark on how to reformulate the moving lemma in order to
make the proof of existence of the intersection product work for
nonsingular quasi-projective varieties (nor just projective ones).