algebra.tex:8151: I found this argument rather nontrivial, so I added it in. Probably it is more complicated than necessary, though; also, the injectivity of the second arrow is not used.
There is one more thing I don't understand. You define exact functors in
categories.tex:2876 using limits and colimits. In algebra.tex, you seem to be
using a different definition: A functor between abelian(?) categories is
left, resp. right, exact if it preserves exactness of sequences of the form 0
-> A -> B -> C, resp. A -> B -> C -> 0. Is the equivalence between these two
definitions clear? Wouldn't it be helpful to make it explicit and emphasize
its use?