Two small changes
1. Removed the comment on line 8218 of algebra.tex because just in the
paragraph preceding it, it was stated that it suffices

2. Try to keep definitions clean.

Comment on exactness: We make the link between exactness of functors
between abelian categories and general categories in Section Tag 010M of
the chapter on homology. As the commutative algebra chapter precedes the
homology chapter, we cannot use this in the commutative algebra chapter.