algebra.tex:8210--8223: Clarifying injectivity argument.
You are right -- it is a straightforward argument; I was mostly confused
by the misconception that I had to use something about tensor products.
It might become clearer if a part of this proof is factored out into a
general criterion for when right-exact functors preserving colimits are
exact; do you see any other applications for such a criterion?