Revised the material in commits Darij Grinberg
Removed the PBW like lemma in commit
b28cd605026f9194c57022c3b301843e52a66f50
Although we should have lemmas like this in the future, it wasn't
necessary for the proof of lemma-finite-finitely-presented-extension.
Instead we moved this lemma later so we can make use of (very
elementary) theory of finite ring maps, thereby reducing to the one
variable case which should be clear to anyone with an undergraduate
course in modern algebra,

Of course this means we *should* add a discussion of the Grobner
basis algorithm and at that point the lemma will be obvious.