Add precise reference to CLO
Thanks to Thomas Smith who writes:

"Reference(and brief explanation) for Proof that an Algebraic Space is a
Scheme if its Reduction is a Scheme.

This result is Corollary 3.1.12 of Conrad and Lieblich's paper
NAGATA COMPACTIFICATION FOR ALGEBRAIC SPACES. The proof proceeds by taking
the reduced space to be affine, thus requiring the space S to be
quasi-compact and seperated. They then apply the central result of the
paper to write S as the inverse limit of a set of algebraic spaces of
finite presentation over the integers. Recognizing that this decomposition
applied to the reduced space results in an affine scheme, they conclude
that each element used in the limit representation of the original space
must be a noetherian algebraic space. Thus, by Theorem 3.3 of Knutson's
Algebriac Spaces, S must be a scheme as all maps from the space into
each element S_i is affine."