Separated grp alg spaces over infinite fields are schemes
The results in this commit beg the following question: Can there be an
algebraic space X over a finite field k such that X_{k(t)} is a scheme,
but X is not a scheme? Which leads to the following similar question:
Can there be a nonrepresentable algebraic space X over k such that both
X_{k'} and X_{k''} are schemes, where k' and k'' are the extensions of
degree 2 and 3 of k?

Weird! Email me if you have an answer...