Better handling invertible modules
Of course an invertible module should be defined as a module such that
tensoring with it is a self-equivalence on the category of modules. This
doesn't change a thing beyong the definition, because we usually are
working on locally ringed spaces or topoi where an invertible module is
the same thing as a locally free module of rank 1.