The picard scheme of a curve
Only the case of a smooth projective curve over an algebraically closed
field. We need it for the chapter on etale cohomology, in order to
esthablish the count of torsion line bundles on a curve over an
algebraically closed field. However, we do not need the result over
arbitrary fields as we will later do Picard stacks in much greater
generality.