Codimension 0 points on algebraic spaces
This is a rather large set of changes all related to different ways of
saying what it means to have a ``generic point'' on an algebraic space.
One particularly nice lemma says that an algebraic space which locally
(in a suitable sense) has finitely many generic points is automatically
a reasonable algebraic space (in particular decent). The proof is the
same as the argument showing that a decent locally Noetherian algebraic
space is quasi-separated.

These types of results may be useful in the future as tools to decide
whether a given algebraic space is decent and/or quasi-separated.