A bit about numerical intersections
On proper schemes over a field you can get numbers using the
Euler characteristic and you can very quickly build a bit of
theory with these. The goal of doing so (in the short term)
was the last lemma of the commit stating how the degrees are
related to degrees of finite morphisms of proper varieties.