Use new terminology and fix indices
Try to be more consistent with convergence and all. It turns out to be
basically impossible, because already the statements are not precise
enough. The trouble is that one needs to always assume the spectral
sequence comes from a filtered complex (somewhere somehow) and then one
is using the definition as now given in homology.tex

Thanks to Dingxin Zhang for pointing out wrong signs in shifts
http://stacks.math.columbia.edu/tag/015X#comment-1479