TAG 04YX
Here's a suggestion for the first lemma of TAG 04YX. It's my first time
trying to help out with the Stacks Project. Please let me know if there
is a more efficient way to do this. (The text below starts at the second
paragraph of the Tag.) In the text preceding the paragraph I wrote $K =
K_{f,x}$ to emphasize that $K$ depends on $f$ and $x$.