Move some lemmes to the correct section
There were two lemmas about the N-1 property in the
section on Nagata rings...

Thanks to Kollar
http://stacks.math.columbia.edu/tag/032G#comment-1539

who says the following:

"The following is a converse to 032G

Claim 1. Let X be a Noetherian, integral scheme whose local rings
are N-1. Then X is N-1 iff X contains a dense, open, normal subset.

Proof. We may assume that X is affine. If Xn→X is finite then is is
an isomorphism over a dense open subset.

Conversely, assume that X∖(g=0) is normal. We apply (Claim 2) to get
a finite partial normalization p:Y→X that is an isomorphism over
X∖(g=0) such that Y is either regular (of dim 1) or has depth ≥2
at all points of (p∗g=0). Thus Y is normal. \qed

Claim 2. Let X be a reduced, Noetherian scheme that is locally N-1.
Let xi∈X be a finite set of points and g a non-zerodivisor. Then
there is a finite partial normalization p:Y→X such that Y is
either regular (of dim 1) or has depth ≥2 at all preimages of
the xi and at all points of (p∗g=0). We can also assume that p
is an isomorphism over X∖(g=0)∪⋃ix¯i.

Proof. Pick any x∈X, take the punctual normalization
(xpn,Xpn)→(x,X) (=maximal partial normalization that is isomorphism
outside x) and then extend it to a partial normalization X′→X that is
an isomorphism over X∖x¯. Using this procedure inductively, we take
care of the first part. (Note: the N-1 condition is inherited by X′.)

To ensure the second part, we first apply this argument to the
generic points of (g=0). Thus we may assume that X is regular at all
generic points of (g=0). The points of (g=0) where X has depth ≤1
correspond to the embedded points of (g=0). We first remove the
generic points of the set of embedded points and then repeat the
argument. \qed"

We decided to not include Claim 2 for the moment.