Improve Lemma 0BLS
With the improved fully faithfulness we can improve this lemma and make
it actually usable. It turns out that the assumptions of the original
form of the lemma actually imply the assumptions of the lemma in the new
form, but it isn't completely trivial that this is so. However, since
the lemma has been online in the Stacks project only for a couple of
days we're going to go ahead and make the change anyways.