Improve lemmas on thickenings and add one
Thanks to Matthew Emerton, Brandon Levin, and Toby Gee who sent these
improvements as well as the new lemma together with other material and
the following comments:

"The write-up begins with some preliminary material related to

http://stacks.math.columbia.edu/tag/05ZJ

In particular, Lemma 1 of the file recalls

http://stacks.math.columbia.edu/tag/05ZJ

and adds the condition of being representable (by schemes) to the
conditions already there.

Lemma 2 is

http://stacks.math.columbia.edu/tag/09ZZ

but adds the conditions of being a monomorphism, or being an
immersion.

Lemma 3 is a variant on Lemma 2, where we remove the hypothesis of
being a finite order thickening, but impose the condition that
both f and f' are locally of finite type."