Algebraic stacks thickening of an algebraic space
Thanks to Matthew Emerton, Toby Gee, and Brandon Levin

Matthew writes:

"Please find attached a PDF file, and the underlying LaTeX file, giving
a proof of the following result, which I couldn't find in the stacks
project, but which might belong there:

An algebraic stack thickening of an algebraic space is again an
algebraic space

and some related results (replace ``algebraic space'' by
``Deligne--Mumford stack'', ``quasi-DM stack'', etc.).

First, apologies if this is already there, and I missed it.

Second, this write-up arose out of discussions with Brandon Levin and
Toby Gee, and reflects contributions of all three of us. Brandon also
told me that he heard of this result from Bhargav Bhatt, but didn't know
of a reference.

<--------snip--------->

It would be great if this could be incorporated into the Stack project.
I think that Toby and I might need things like this in our work at some
point, and I think maybe Brandon does as well, and having the stacks
project as a reference for it would be very convenient!

Cheers,

Matt"